Move the auth components into the auth folder for simplicity
