+++ packages/react/spec/components/auth/SignedIn.spec.tsx
@@ -1,7 +1,7 @@
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
 import React from "react";
-import { SignedIn } from "../../../src/components/auth/SignedIn.js";
+import { SignedIn } from "../../../src/auth/SignedIn.js";
 import { superAuthApi } from "../../apis.js";
 import { MockClientWrapper } from "../../testWrappers.js";
 import { expectMockDeletedUser, expectMockSignedInUser, expectMockSignedOutUser } from "../../utils.js";
