+++ packages/react/spec/components/auth/SignedInOrRedirect.spec.tsx
@@ -1,7 +1,7 @@
 import "@testing-library/jest-dom";
 import { render } from "@testing-library/react";
 import React from "react";
-import { SignedInOrRedirect } from "../../../src/components/auth/SignedInOrRedirect.js";
+import { SignedInOrRedirect } from "../../../src/auth/SignedInOrRedirect.js";
 import { superAuthApi } from "../../apis.js";
 import { MockClientWrapper } from "../../testWrappers.js";
 import { expectMockDeletedUser, expectMockSignedInUser, expectMockSignedOutUser } from "../../utils.js";
